+++ CHANGELOG.md
@@ -8,6 +8,10 @@ From v1.0.0 and on, the project will adherence strictly to Semantic Versioning.
 
 
 ## [Unreleased]
+### Added
+- `co.Lambda` module.
+- `co.Add` module.
+- `co.Multiply` module.
 
 
 ## [0.8.1]